Started to update to last rustc version (removing box closure)

+++ src/gdk/types.rs
@@ -18,10 +18,10 @@ use libc::{c_void};
 #[repr(C)]
 #[derive(Copy)]
 pub struct Rectangle { //FIXME should be just an alias to cairo_rectangle_int_t
-    x: int,
-    y: int,
-    width: int,
-    height: int
+    x: isize,
+    y: isize,
+    width: isize,
+    height: isize
 }
 
 pub type Atom = *mut c_void;
